Refactoring and handling of unused
