+++ src/main/java/com/joshcummings/codeplay/terracotta/http/HttpSupport.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.joshcummings.codeplay.terracotta.testng;
+package com.joshcummings.codeplay.terracotta.http;
 
 import java.io.IOException;
