+++ src/test/java/com/joshcummings/codeplay/terracotta/AbstractEmbeddedTomcatSeleniumTest.java
@@ -18,7 +18,7 @@ package com.joshcummings.codeplay.terracotta;
 import java.time.Duration;
 import java.util.function.Function;
 
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 import com.joshcummings.codeplay.terracotta.testng.ProxySupport;
 import com.joshcummings.codeplay.terracotta.testng.SeleniumSupport;
 import com.joshcummings.codeplay.terracotta.testng.TestConstants;
