+++ src/vulnerability-test/java/com/joshcummings/codeplay/terracotta/AbstractEmbeddedTomcatTest.java
@@ -1,7 +1,7 @@
 package com.joshcummings.codeplay.terracotta;
 
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 import com.joshcummings.codeplay.terracotta.testng.DockerSupport;
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
 import com.joshcummings.codeplay.terracotta.testng.TomcatSupport;
 import org.testng.ITestContext;
 import org.testng.annotations.AfterTest;
