+++ src/vulnerability-test/java/com/joshcummings/codeplay/terracotta/BankTransferFunctionalTest.java
@@ -18,11 +18,11 @@ package com.joshcummings.codeplay.terracotta;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 import com.joshcummings.codeplay.terracotta.model.Account;
 import com.joshcummings.codeplay.terracotta.model.Client;
 import com.joshcummings.codeplay.terracotta.service.AccountService;
 import com.joshcummings.codeplay.terracotta.service.ClientService;
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
 import com.joshcummings.codeplay.terracotta.testng.TomcatSupport;
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.RequestBuilder;
