+++ src/vulnerability-test/java/com/joshcummings/codeplay/terracotta/ChangePasswordFunctionalTest.java
@@ -15,11 +15,11 @@
  */
 package com.joshcummings.codeplay.terracotta;
 
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 import com.joshcummings.codeplay.terracotta.model.Transaction;
 import com.joshcummings.codeplay.terracotta.model.User;
 import com.joshcummings.codeplay.terracotta.service.TransactionService;
 import com.joshcummings.codeplay.terracotta.service.UserService;
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
 import com.joshcummings.codeplay.terracotta.testng.TomcatSupport;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.testng.Assert;
