+++ src/vulnerability-test/java/com/joshcummings/codeplay/terracotta/ForgotPasswordFunctionalTest.java
@@ -17,11 +17,11 @@ package com.joshcummings.codeplay.terracotta;
 
 import java.util.Collection;
 
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 import com.joshcummings.codeplay.terracotta.model.Transaction;
 import com.joshcummings.codeplay.terracotta.model.User;
 import com.joshcummings.codeplay.terracotta.service.TransactionService;
 import com.joshcummings.codeplay.terracotta.service.UserService;
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
 import com.joshcummings.codeplay.terracotta.testng.TomcatSupport;
 import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
