+++ src/vulnerability-test/java/com/joshcummings/codeplay/terracotta/LoginFunctionalTest.java
@@ -192,6 +192,7 @@ public class LoginFunctionalTest extends AbstractEmbeddedTomcatSeleniumTest {
 		Assert.assertTrue(content.contains("Something about the login provided is incorrect."));
 	}
 
+	@SuppressWarnings("unused")
 	private Map.Entry<String, String> attemptLogin(String username) {
 		String content = http.postForContent(
 				post("/login")
