+++ src/vulnerability-test/java/com/joshcummings/codeplay/terracotta/app/ContentParsingFilterVTests.java
@@ -214,7 +214,9 @@ public class ContentParsingFilterVTests extends AbstractEmbeddedTomcatTest {
 	@Test
 	public void testJsonDeserializationGadget() throws Exception {
 		ObjectMapper o = new ObjectMapper();
-		o.enableDefaultTyping(); // scary
+		// Replace deprecated call with modern equivalent
+		o.activateDefaultTyping(o.getPolymorphicTypeValidator()); 
+		
 		Person in = new Person(new EvilObject(false), Arrays.asList("some", "addresses"));
 		byte[] ser = jsonSerialize(in, o);
 		System.out.println(new String(ser));
@@ -258,9 +260,15 @@ public class ContentParsingFilterVTests extends AbstractEmbeddedTomcatTest {
 		}
 		return baos.toByteArray();
 	}
+
+	// Fix type safety warning in javaDeserialize method
 	private <T> T javaDeserialize(InputStream is, Class<T> clazz) throws IOException, ClassNotFoundException {
 		try (ObjectInputStream ois = new ObjectInputStream(is)) {
-			return (T) ois.readObject();
+			Object obj = ois.readObject();
+			if (clazz.isInstance(obj)) {
+				return clazz.cast(obj);
+			}
+			throw new ClassNotFoundException("Object is not of expected type " + clazz.getName());
 		}
 	}
 
@@ -268,19 +276,19 @@ public class ContentParsingFilterVTests extends AbstractEmbeddedTomcatTest {
 		private List<String> addresses;
 		private Object details;
 
-		public Person() {}
-
 		public Person(Object details, List<String> addresses) {
 			if (addresses == null || addresses.size() < 2)
 				throw new IllegalArgumentException("bad argument");
 			this.addresses = new ArrayList<>(addresses);
 			this.details = details;
 		}
-
+		
+		@SuppressWarnings("unused")
 		public List<String> getAddresses() {
 			return addresses;
 		}
 
+		@SuppressWarnings("unused")
 		public Object getDetails() {
 			return details;
 		}
@@ -294,6 +302,7 @@ public class ContentParsingFilterVTests extends AbstractEmbeddedTomcatTest {
 
 	@JsonRootName("evil")
 	private static class EvilObject implements Serializable {
+		@SuppressWarnings("unused")
 		public EvilObject() {
 			fail("attempted to deserialize EvilObject");
 		}
