+++ src/vulnerability-test/java/com/joshcummings/codeplay/tools/BruteForceTester.java
@@ -15,9 +15,10 @@
  */
 package com.joshcummings.codeplay.tools;
 
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
+
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
