+++ src/vulnerability-test/java/com/joshcummings/codeplay/tools/EnumerationTester.java
@@ -15,9 +15,10 @@
  */
 package com.joshcummings.codeplay.tools;
 
-import com.joshcummings.codeplay.terracotta.testng.HttpSupport;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
+
+import com.joshcummings.codeplay.terracotta.http.HttpSupport;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
